BROKEN SH48 EXAMPLE

+++ src/piston_sensor.rs
@@ -23,11 +23,11 @@ impl WavefrontSensor for PistonSensor {
     fn calibrate(&mut self, src: &mut crate::Source, _threshold: f64) {
         self.piston_refs = src.segment_piston();
     }
-    fn data(&mut self) -> Vec<f32> {
+    fn data(&mut self) -> Vec<f64> {
         self.pistons
             .iter()
             .zip(&self.piston_refs)
-            .map(|(x, x0)| (*x - *x0) as f32)
+            .map(|(x, x0)| (*x - *x0))
             .collect()
     }
 }
